Unknown Metric Filtering Enhancements
Adds a field to metadata.yaml called `sendUnknown`, which, if true, will
cause the monitor built-in filtering logic to send any metrics that are
not specified in the metadata.yaml metrics section.  This is useful for
monitors like JMX where the metric names are defined by complex
user-provided configuration and cannot be predicted ahead of time. This
replaces the half-implemented `metricsExhaustive` field that never
really did anything.

Now all unknown metrics for monitors which do not have `sendUnknown:
true` will be dropped by default.  Previously all unknown metrics from
all monitors would be let through.

Also enhances the built-in filtering to automatically send everything if
there are no metrics defined.  This is largely redundant with the
existing `sendAll` flag but it provides a good safeguard to avoid
confusing filtering behavior if the `sendAll` flag is omitted on
accident.